requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.15.1
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-password-validators==1.7.1
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
